+++ packages/terra-form-input/tests/wdio/form-input-spec.js
@@ -1,28 +1,28 @@
 Terra.describeViewports('Form-Input', ['tiny', 'large'], () => {
   it('should display valid Form Input', () => {
     browser.url('/raw/tests/terra-form-input/form-input/default-form-input');
-    browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
     Terra.validates.element('valid');
   });
 
   it('should display invalid Form Input', () => {
     browser.url('/raw/tests/terra-form-input/form-input/invalid-form-input');
-    browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
     Terra.validates.element('incomplete');
   });
 
   it('should display incomplete Form Input', () => {
     browser.url('/raw/tests/terra-form-input/form-input/incomplete-form-input');
-    browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
     Terra.validates.element('invalid');
   });
 
   it('should display invalid Form Input with incomplete', () => {
     browser.url('/raw/tests/terra-form-input/form-input/invalid-with-incomplete-form-input');
-    browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
     Terra.validates.element('invalid with incomplete');
   });
@@ -30,21 +30,21 @@ Terra.describeViewports('Form-Input', ['tiny', 'large'], () => {
   describe('Enabled', () => {
     it('should display enabled Form Input non-selected', () => {
       browser.url('/raw/tests/terra-form-input/form-input/default-form-input');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
       Terra.validates.element('non-selected');
     });
 
     it('should display enabled Form Input hovered', () => {
       browser.url('/raw/tests/terra-form-input/form-input/default-form-input');
-      browser.moveToObject('input');
+      $('input').moveTo();
 
       Terra.validates.element('hovered');
     });
 
     it('should display enabled Form Input with value', () => {
       browser.url('/raw/tests/terra-form-input/form-input/value-form-input');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
       Terra.validates.element('value');
     });
@@ -52,7 +52,7 @@ Terra.describeViewports('Form-Input', ['tiny', 'large'], () => {
     it('should display enabled Form Input keyboard-focused', () => {
       browser.url('/raw/tests/terra-form-input/form-input/default-form-input');
       browser.keys('Tab');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
       Terra.validates.element('keyboard-focused');
     });
@@ -61,72 +61,72 @@ Terra.describeViewports('Form-Input', ['tiny', 'large'], () => {
   describe('Disabled', () => {
     it('should display disabled form Input non-selected', () => {
       browser.url('/raw/tests/terra-form-input/form-input/disabled-form-input');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('non-selected');
+      Terra.validates.element('disabled non-selected');
     });
 
     it('should display disabled Form Input with value', () => {
       browser.url('/raw/tests/terra-form-input/form-input/disabled-value-input');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('value');
+      Terra.validates.element('disabled value');
     });
 
     it('should display disabled Form Input hovered', () => {
       browser.url('/raw/tests/terra-form-input/form-input/disabled-form-input');
-      browser.moveToObject('input');
+      $('input').moveTo();
 
-      Terra.validates.element('hovered');
+      Terra.validates.element('disabled hovered');
     });
 
     it('should display disabled Form Input keyboard-focused', () => {
       before(() => {
         browser.url('/raw/tests/terra-form-input/form-input/disabled-form-input');
         browser.keys('Tab');
-        browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+        $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
       });
 
-      Terra.validates.element('keyboard-focused');
+      Terra.validates.element('disabled keyboard-focused');
     });
   });
 
   describe('InputField', () => {
     it('should display resized InputField', () => {
       browser.url('/raw/tests/terra-form-input/form-input/input-field-width');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('resized');
+      Terra.validates.element('input field resized');
     });
 
     it('should display valid InputField', () => {
       browser.url('/raw/tests/terra-form-input/form-input/input-field');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('valid');
+      Terra.validates.element('input field valid');
     });
 
     it('should display invalid InputField', () => {
       browser.url('/raw/tests/terra-form-input/form-input/input-field');
-      browser.click('#validity-toggle');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#validity-toggle').click();
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('invalid');
+      Terra.validates.element('input field invalid');
     });
 
     it('should display incomplete InputField', () => {
       browser.url('/raw/tests/terra-form-input/form-input/input-field');
-      browser.click('#incomplete-toggle');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#incomplete-toggle').click();
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('incomplete');
+      Terra.validates.element('input field incomplete');
     });
 
     it('should display disabled InputField', () => {
       browser.url('/raw/tests/terra-form-input/form-input/disabled-input-field');
-      browser.moveToObject('#root', 0, 700); // add to ensure consistent hover styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // add to ensure consistent hover styles
 
-      Terra.validates.element('disabled');
+      Terra.validates.element('input field disabled');
     });
   });
 });
@@ -140,7 +140,7 @@ Terra.describeViewports('Form-Input', ['large'], () => {
     });
 
     it('should trigger onInput when programmatically setting value', () => {
-      browser.click('#input-set-text-button');
+      $('#input-set-text-button').click();
 
       Terra.validates.element('programmatically');
     });