+++ packages/terra-form-select/tests/wdio/select-field-spec.js
@@ -3,17 +3,17 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/select-field'));
 
     it('should select an option by click', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('select with click');
     });
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('#select-field:last-child');
+      $('#select-field:last-child').click();
 
       Terra.validates.element('open-dropdown', { selector: '#root' });
     });
 
     it('should select the first option', () => {
-      browser.click('#terra-select-option-xSmall');
+      $('#terra-select-option-xSmall').click();
 
       Terra.validates.element('selected-option');
     });
@@ -41,20 +41,20 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/select-field'));
 
     it('should select an option by pressing enter', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('select with enter');
     });
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('#select-field:last-child');
+      $('#select-field:last-child').click();
 
-      Terra.validates.element('open-dropdown', { selector: '#root' });
+      Terra.validates.element('open dropdown after enter', { selector: '#root' });
     });
 
     it('should select the first option by pressing enter', () => {
       browser.keys('ArrowUp');
       browser.keys('Enter');
 
-      Terra.validates.element('selected-option');
+      Terra.validates.element('selected option with enter');
     });
   });
 
@@ -62,35 +62,35 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/max-selection'));
 
     it('should select the maximum number of items allowed', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('select with max item selections');
     });
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('[data-terra-select-combobox]');
+      $('[data-terra-select-combobox]').click();
 
-      Terra.validates.element('open-dropdown', { selector: '#root' });
+      Terra.validates.element('open dropdown for max selection', { selector: '#root' });
     });
 
     it('should select the first option', () => {
-      browser.click('#terra-select-option-blue');
+      $('#terra-select-option-blue').click();
 
       Terra.validates.element('selected-first-option', { selector: '#root' });
     });
 
     it('should select the second option', () => {
-      browser.click('#terra-select-option-cyan');
+      $('#terra-select-option-cyan').click();
 
       Terra.validates.element('max-selection-count-reached', { selector: '#root' });
     });
 
     it('should unselect the second option', () => {
-      browser.click('#terra-select-option-cyan');
+      $('#terra-select-option-cyan').click();
 
       Terra.validates.element('unselect-second-option', { selector: '#root' });
     });
 
     it('should select the second option again', () => {
-      browser.click('#terra-select-option-green');
+      $('#terra-select-option-green').click();
 
       Terra.validates.element('max-selection-count-reached-again', { selector: '#root' });
     });
@@ -117,20 +117,20 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     });
 
     it('should show the max selection option on search', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('max selection on search');
     });
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('[data-terra-select-combobox]');
+      $('[data-terra-select-combobox]').click();
 
-      Terra.validates.element('open-dropdown', { selector: '#root' });
+      Terra.validates.element('[max selection option on search]open dropdown', { selector: '#root' });
     });
 
     it('should select the first option', () => {
-      browser.click('#terra-select-option-blue');
-      browser.click('#terra-select-option-cyan');
+      $('#terra-select-option-blue').click();
+      $('#terra-select-option-cyan').click();
 
-      Terra.validates.element('max-selection-count-reached', { selector: '#root' });
+      Terra.validates.element('max selection count reached on search', { selector: '#root' });
     });
 
     it('should show the max selection option', () => {
@@ -144,17 +144,17 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/uncontrolled-search'));
 
     it('should be accessible with no results', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('no results');
     });
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('#search:last-child');
+      $('#search:last-child').click();
     });
 
     it('select should enter a free text entry', () => {
       browser.keys(['T', 'a', 'g']);
 
-      Terra.validates.element('open-dropdown', { selector: '#root' });
+      Terra.validates.element('free text entry', { selector: '#root' });
     });
   });
 
@@ -162,11 +162,11 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/combobox-field-max-height'));
 
     it('should open the Combobox field and honor the set max height', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('combobox max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('#comboboxField:last-child');
+      $('#comboboxField:last-child').click();
 
       Terra.validates.element('combobox-field-dropdown-max-height', { selector: '#root' });
     });
@@ -175,12 +175,12 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
   describe('should open the Multi select field and honor the set max height', () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/multi-select-field-maxheight'));
 
-    it('should open the Multi select field and honor the set max heigh', () => {
-      Terra.validates.element('default');
+    it('should open the Multi select field and honor the set max height', () => {
+      Terra.validates.element('multi select max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('[data-terra-form-select-toggle]');
+      $('[data-terra-form-select-toggle]').click();
 
       Terra.validates.element('multi-select-dropdown-max-height', { selector: '#root' });
     });
@@ -190,11 +190,11 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/search-select-field-maxheight'));
 
     it('should open the search select and honor the set max height', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('search select max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('#searchSelectField:last-child');
+      $('#searchSelectField:last-child').click();
 
       Terra.validates.element('search-select-dropdown-max-height', { selector: '#root' });
     });
@@ -204,11 +204,11 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/select-field-maxheight'));
 
     it('should open the Select field and honor the set max height', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('select field max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('#selectField:last-child');
+      $('#selectField:last-child').click();
 
       Terra.validates.element('select-field-dropdown-max-height', { selector: '#root' });
     });
@@ -218,11 +218,11 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/single-select-field-maxheight'));
 
     it('should open the Single select field and honor the set max height', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('single select max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('#singleSelectField:last-child');
+      $('#singleSelectField:last-child').click();
 
       Terra.validates.element('single-select-dropdown-max-height', { selector: '#root' });
     });
@@ -232,11 +232,11 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     before(() => browser.url('/raw/tests/terra-form-select/form-select/tag-select-field-maxheight'));
 
     it('should open the Tag select field and honor the set max height', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('tag select max height');
     });
 
     it('default should open the dropdown by clicking the toggle', () => {
-      browser.click('[data-terra-form-select-toggle]');
+      $('[data-terra-form-select-toggle]').click();
 
       Terra.validates.element('tag-select-dropdown-max-height', { selector: '#root' });
     });