+++ packages/terra-action-footer/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated Jest command
+
 ## 2.58.0 - (February 2, 2021)
 
 * Changed
@@ -11,7 +14,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 2.56.0 - (December 8, 2020)
 
 * Changed